Split stream, settings, and cloud session modules
- Extract input, overlay, account, action, page, protocol, and state logic
- Refresh repository documentation and remove obsolete assets and build scripts
