--- .github/FUNDING.yml
@@ -1,2 +1,0 @@
-custom:
-  - https://nowpayments.io/payment/?iid=5035597688
